+++ .env.example
@@ -1,4 +1,5 @@
 REACT_APP_NODE_ENV=development|production
 REACT_APP_GOOGLE_ANALYTICS_ID='UA-XXXXXXX'
 REACT_APP_WORKSHOPCHALLENGE_API_ENDPOINT=
+REACT_APP_CHALLENGE_API_ENDPOINT=
 REACT_APP_MONGODB_URI=mongodb://localhost/discover-game
